requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.2
djangorestframework==3.14.0
httpie==3.2.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
Pygments==2.17.2
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
